# internal/handler/static/openapi.yaml
openapi: 3.0.3
info:
  title: Simple Host API
  description: >-
    Static website hosting with a light per-site backend. Deploy a site with one
    API call — inline JSON files (LLM-friendly) or a tar.gz/zip archive — and each
    site gets shared JSON state (with atomic ops), append-only collections. Sign-in is a 6-digit email code or Google that
    returns an API key, sent as the
    `X-API-Key` header. Hosted-page sign-in uses the same account but a
    site-scoped session cookie on the site's own address — never an API key.
    Every site lives at its own address, `https://<site>.<handle>.simple-host.app/`
    (its own browser origin); `https://<handle>.simple-host.app/` lists the
    person's public sites. For a brand-new account a site may briefly live at
    `https://<handle>.simple-host.app/<site>/` until its certificate is issued:
    use the `site_url` the API returns, never a composed one.
    Old `<handle>.simple-host.app/<site>/` links redirect to the site's address,
    as do old `sites.simple-host.app/<handle>/<site>/` ones.
    Visitors sign in with Google or an emailed code on the site's own address (a
    sign-in covers that site only); every save from a page needs a signed-in visitor. Agents save with the API key
    (or the connector) through `https://simple-host.app/v1/...`.
  version: 2.0.0
servers:
  - url: https://simple-host.app
...
paths:
...
  /v1/sites/{sitename}/domain:
    post:
      operationId: postSitesBySitenameDomain
      summary: Give this site a nicer address (free <name>.simple-host.app or a custom domain)
      description: >-
        Optional. Every site already lives at its own address
        `https://<site>.<handle>.simple-host.app/`, with sign-in and private lists.
        This gives it a nicer address; the site moves there and its old address redirects.
        A free address: send `{"domain": "<name>.simple-host.app"}` (one DNS label:
        lowercase letters, digits and hyphens, not at the start or end, up to 63
        characters). It is active at once with no DNS step — the answer has
        `status: active` and `verified_at`, and no `dns`. First come, first served:
        409 domain_taken if another site holds it or it is already someone's handle
        (handles and free names share one namespace); reserved names (www, sites, api,
        admin, mail, cname and the platform's own hosts) get 400 name_reserved;
        anything else malformed gets 400 invalid_name. Once claimed it behaves
        exactly like a custom domain: the site is served at its root with a
        same-origin /v1 API and visitor sign-in, its person-address URL 302s there, and
        private collections take submissions there. A site has one
        own address: claiming one replaces a custom domain and vice versa. A free
        name the site lets go (switching, DELETE, or deleting the site) stays with
        it: it redirects to the site's current address and nobody else can claim it.
        A custom domain: bind ONE domain to the site (status starts as `pending`). Returns
        the DNS record the owner must create (CNAME for a subdomain, A for an apex).
        Once DNS points here, its certificate is requested automatically
        (`certificate_status`: pending → issuing → live, or failed with the reason in
        `last_error`, retried) and the domain is served at its root, usually within
        minutes. If the site already has a working own address (a free name or an
        earlier domain), it keeps serving there (`previous_domain`) until the new
        domain is verified; then the old address redirects to it. Binding the
        site's current working domain again changes nothing; binding its
        `previous_domain` drops the pending one. Until first verified, a binding is
        provisional: another site can take it over, and it expires 24 hours after
        binding unless DNS already points here. A verified domain stays exclusive
        while it works; if it fails every check for 24 hours the owner is emailed,
        and after 72 hours it stops being the site's address (the site's own
        address serves again) and can be connected afresh by whoever holds it.
        Hijacking the platform zone is rejected. Owner only (API key).
      tags: [Custom domains]
      security: [{ apiKey: [] }]
      parameters: [ { $ref: '#/components/parameters/sitename' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [domain]
              properties:
                domain: { type: string, example: www.example.com }
      responses:
        '200':
          description: Domain bound — pending DNS for a custom domain, active at once for a free <name>.simple-host.app address
          content:
            application/json:
              schema:
                type: object
                properties:
                  took_over_from: { type: string, description: Previous holder as handle/site; included only on takeover of an unproven binding, example: alice/blog }
                  domain: { type: string, example: www.example.com }
                  status: { type: string, enum: [pending, active, error], example: pending }
                  bound_at: { type: string, format: date-time }
                  verified_at: { type: string, format: date-time, nullable: true, description: Set at once for a free <name>.simple-host.app address }
                  expires_at: { type: string, format: date-time, description: bound_at plus 24 hours; included only while pending and never verified }
                  certificate_status: { type: string, enum: [pending, issuing, live, failed], description: "The custom domain's certificate: pending (DNS not pointed here yet), issuing, live, or failed (why in last_error; retried). Absent for a free <name>.simple-host.app address." }
                  previous_domain: { type: string, description: "The site's earlier own address, still serving it until this domain is verified (then it redirects here). Present only while the new domain is pending.", example: clay-studio.simple-host.app }
                  dns:
                    description: The record to add at the registrar; absent for a free <name>.simple-host.app address
                    type: object
                    properties:
                      type: { type: string, example: CNAME }
                      host: { type: string, example: www.example.com }
                      value: { type: string, example: cname.simple-host.app }
        '400':
          description: Invalid domain. For a <name>.simple-host.app address, `code` is name_reserved or invalid_name.
          content:
            application/json:
              schema: { type: object, properties: { error: { type: string }, code: { type: string, enum: [name_reserved, invalid_name] } } }
              examples:
                reserved: { value: { error: "that name is reserved; pick another", code: name_reserved } }
                invalid: { value: { error: "the name may use lowercase letters, numbers and hyphens (not at the start or end), up to 63 characters", code: invalid_name } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: Domain has been verified by another site at least once, or the <name>.simple-host.app address is held by another site or is someone's handle
          content:
            application/json:
              schema:
                type: object
                required: [error, code]
                additionalProperties: false
                properties:
                  error: { type: string, enum: [domain is connected to another site, "that address is taken by another site; pick another name"] }
                  code: { type: string, enum: [domain_taken] }
              example: { error: domain is connected to another site, code: domain_taken }
    get:
      operationId: getSitesBySitenameDomain
      summary: Get the custom domain binding for this site
      description: >-
        Returns the bound domain and status, or `{domain:null,status:null}` when
        none is set. Includes the DNS record to add when a custom domain is bound,
        the certificate's progress, the earlier address still serving while the
        domain is pending, and `failing_since` when a verified domain has stopped
        passing its checks. Owner only (API key).
      tags: [Custom domains]
      security: [{ apiKey: [] }]
      parameters: [ { $ref: '#/components/parameters/sitename' } ]
      responses:
        '200':
          description: Current domain binding
          content:
            application/json:
              schema:
                type: object
                properties:
                  domain: { type: string, nullable: true, example: www.example.com }
                  status: { type: string, nullable: true, enum: [pending, active, error] }
                  bound_at: { type: string, format: date-time, description: Binding timestamp stored as domain_bound_at; reset on bind }
                  expires_at: { type: string, format: date-time, description: bound_at plus 24 hours; included only while pending and never verified }
                  verified_at: { type: string, format: date-time, nullable: true }
                  last_error: { type: string, description: Why the domain is not active, from the latest check }
                  certificate_status: { type: string, enum: [pending, issuing, live, failed], description: "pending (DNS not pointed here yet), issuing, live, or failed (why in last_error; retried). Absent for a free <name>.simple-host.app address." }
                  previous_domain: { type: string, description: "The site's earlier own address, still serving it until this domain is verified. Present only while the domain is pending." }
                  failing_since: { type: string, format: date-time, description: "A verified domain that stopped passing its checks: the owner is emailed after 24 hours; after 72 hours it stops being the site's address." }
                  dns:
                    type: object
                    properties:
                      type: { type: string, example: CNAME }
                      host: { type: string }
                      value: { type: string }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      operationId: deleteSitesBySitenameDomain
      summary: Disconnect this site's domain or free address
      description: >-
        Disconnects the domain GET shows. If it was still pending and the site kept
        serving at an earlier address (`previous_domain`), that address becomes the
        site's domain again; otherwise the site is served at its own address.
        A disconnected free <name>.simple-host.app stays with the site and redirects
        to its current address; links to a disconnected custom domain stop working.
        Owner only (API key).
      tags: [Custom domains]
      security: [{ apiKey: [] }]
      parameters: [ { $ref: '#/components/parameters/sitename' } ]
      responses:
        '204': { description: Domain unbound }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /v1/sites/{sitename}/domain/check:
    post:
      operationId: checkSitesBySitenameDomain
      summary: Check this site's domain again now
      description: >-
        Re-proves the site's own domain at once (it resolves here and serves the
        site over HTTPS), instead of waiting for the background check every two
        minutes, then answers exactly like GET .../domain with the new status and
        `last_error`. A free <name>.simple-host.app address needs no proof and is
        answered as is. Owner only (API key); rate-limited per IP.
      tags: [Custom domains]
      security: [{ apiKey: [] }]
      parameters: [ { $ref: '#/components/parameters/sitename' } ]
      responses:
        '200':
          description: The domain binding after the check (same shape as GET .../domain)
          content:
            application/json:
              schema:
                type: object
                properties:
                  domain: { type: string, example: www.example.com }
                  status: { type: string, enum: [pending, active, error] }
                  last_error: { type: string, example: "resolves to 203.0.113.9, not to this server" }
                  expires_at: { type: string, format: date-time }
                  verified_at: { type: string, format: date-time, nullable: true }
                  dns: { type: object, properties: { type: { type: string }, host: { type: string }, value: { type: string } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404':
          description: No such site, or the site has no domain connected (no_domain)
          content: { application/json: { schema: { $ref: '#/components/schemas/Error' }, example: { error: this site has no domain connected, code: no_domain } } }
        '429': { $ref: '#/components/responses/TooManyRequests' }
...
components:
  securitySchemes:
    apiKey: { type: apiKey, in: header, name: X-API-Key }
    connector:
      type: oauth2
      description: >-
        Per-person OAuth 2.1 (authorization code; PKCE S256 required for every self-registered
        client). An access token sent as `Authorization: Bearer` acts exactly as that person's
        X-API-Key on every route here. Discovery at /.well-known/oauth-authorization-server.
      flows:
        authorizationCode:
          authorizationUrl: https://simple-host.app/oauth/authorize
          tokenUrl: https://simple-host.app/oauth/token
          refreshUrl: https://simple-host.app/oauth/token
          scopes:
            sites: Publish and manage your sites and their data
  parameters:
...
    sitename: { name: sitename, in: path, required: true, schema: { type: string }, example: my-cool-site }
...
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string, description: What went wrong, in plain words. The wording may change; branch on `code`. }
        code:
          type: string
          description: >-
            A stable snake_case name for the refusal, present where one status
            has more than one meaning. site_exists (409: this account already
            has a site of that name), domain_taken (409: the address belongs to
            another site or person), invalid_name (400: not 1-63 lowercase
            letters, digits and hyphens), name_reserved (400: reserved by the
            platform), invalid_domain (400: not a domain that can be connected),
            site_quota_reached (403), append_only (409: public list items cannot
            be edited; the owner may still delete them), custom_domain_required (409: the site needs
            its own address first), not_an_object (409: the saved value is not a
            JSON object), private_visitor_only and private_needs_own_domain
            (403), use_custom_domain and visitor_auth_required (401),
            missing_api_key, wrong_auth_header (sent Authorization instead
            of X-API-Key) and invalid_api_key (unknown, revoked or signed-out
            key) (401), csrf_required (403), recently_deleted (409: the name
            belongs to a site in Recently deleted), site_not_deleted (409:
            restore of a live site), confirm_required (400),
            not_found (404, private collections), export_link_invalid (404:
            the download link expired or is not valid), site_suspended (403:
            the operator has taken the site down; changes and public reads of
            its data are refused until it is restored, `reason` says why),
            account_suspended (403: the operator has suspended this account;
            its keys, connected apps and sign-in are refused, `reason` says
            why). New codes may be added.
          example: domain_taken
  responses:
...
    Unauthorized: { description: "Missing or invalid API key — send the X-API-Key header, not Authorization Bearer. The body's code is missing_api_key, wrong_auth_header or invalid_api_key.", content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...
    NotFound: { description: Not found, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...
    TooManyRequests: { description: Rate limited, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...